+++ src/commands/user.rs
@@ -6,7 +6,7 @@
 use clap::{Arg, ArgMatches, Command};
 
 use crate::commands::params::{
-    format_parameter, format_pretty_parameter, format_with_headers_parameter,
+    format_parameter, format_pretty_parameter, format_with_headers_parameter, tenant_parameter,
 };
 
 /// Define the user command and its subcommands
@@ -19,6 +19,7 @@ pub fn user_command() -> Command {
             Command::new("list")
                 .about("List users in the current tenant")
                 .visible_alias("ls")
+                .arg(tenant_parameter())
                 .arg(format_parameter().value_parser(["json", "csv"]))
                 .arg(format_pretty_parameter())
                 .arg(format_with_headers_parameter()),
@@ -33,6 +34,7 @@ pub fn user_command() -> Command {
                         .num_args(1)
                         .value_parser(clap::value_parser!(String)),
                 )
+                .arg(tenant_parameter())
                 .arg(format_parameter().value_parser(["json", "csv"]))
                 .arg(format_pretty_parameter())
                 .arg(format_with_headers_parameter()),